Added the notebook of the First part
This is the notebook of the First part about the blackjack. I put the whole description of the game an explanation of the environment, action space, end conditions and reward. I start defining the Naive agent but still have nothing useful yet. I let the rest of the code from the other notebook